Minor tweaks.
We've found that students, for the most part, have poor time management skills and unless provided with a clear framework have a tendency to under-deliver. To mitigate against this, we've moved towards providing guidelines, giving a clear indication of our weekly expectations. We're trialling this in 2013/14 and will leave comments once we've assessed the work at the end of the semester.